+++ swiftui-project/View/ProductDetailsView.swift
@@ -22,7 +22,7 @@ struct ProductDetailsView: View {
                     Button(action: {}, label: {
                         Image(systemName: "cart.fill")
                             .font(.title2)
-                            .foregroundColor(Color.black)
+                            .foregroundColor(.black)
                     })
                 }
                 .padding(.horizontal)
@@ -42,9 +42,8 @@ struct ProductDetailsView: View {
                                 .foregroundColor(.black)
                             
                         }
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 10)
-                        .background(Color.white)
+                        .padding(10)
+                        .background(.white)
                         .cornerRadius(26)
                         .padding(.vertical,15)
                         .padding(.horizontal, 30)
@@ -62,20 +61,25 @@ struct ProductDetailsView: View {
                             .font(.headline)
                             .bold()
                         Image(systemName: "star.fill")
-                            .font(.subheadline)
-                            .foregroundColor(Color(UIColor(named: "TextColor")!))
+                            .resizable()
+                            .frame(width: 12,height: 12)
+                            .foregroundColor(Color("TextColor"))
                         Image(systemName: "star.fill")
-                            .font(.subheadline)
-                            .foregroundColor(Color(UIColor(named: "TextColor")!))
+                            .resizable()
+                            .frame(width: 12,height: 12)
+                            .foregroundColor(Color("TextColor"))
                         Image(systemName: "star.fill")
-                            .font(.subheadline)
-                            .foregroundColor(Color(UIColor(named: "TextColor")!))
+                            .resizable()
+                            .frame(width: 12,height: 12)
+                            .foregroundColor(Color("TextColor"))
                         Image(systemName: "star.fill")
-                            .font(.subheadline)
-                            .foregroundColor(Color(UIColor(named: "TextColor")!))
+                            .resizable()
+                            .frame(width: 12,height: 12)
+                            .foregroundColor(Color("TextColor"))
                         Image(systemName: "star.fill")
-                            .font(.subheadline)
-                            .foregroundColor(Color(UIColor(named: "TextColor")!))
+                            .resizable()
+                            .frame(width: 12,height: 12)
+                            .foregroundColor(Color("TextColor"))
                     }
                     
                     Text("Category: Home")
@@ -100,7 +104,7 @@ struct ProductDetailsView: View {
                     Text("Add to cart")
                         .foregroundColor(.white)
                         .frame(width: 340, height: 46)
-                        .background(Color(UIColor(named: "TextColor")!))
+                        .background(Color("TextColor"))
                         .cornerRadius(26)
                 })
                 .padding(.horizontal,25)
